refactor: hoist shared contracts to src/shared, and make two boundaries absolute
`convex/**` imported `src/app/*/validation` in seven files and the publisher
Worker imported `src/app/capture/publisher-diagnostics` in ten. Both were
described as the codebase's normal shape. They are better read as
mis-filing: every one of those modules is pure — the validators import
nothing but `zod`, and the publisher diagnostics and fonts import nothing at
all — so nothing was being dragged into the Convex or Worker bundles. They
were shared contracts parked on one side, with two separate deploy artifacts
reaching in to get them.

Moved to `src/shared/`, where four asset-publishing modules already lived
(the same hoist, started and left half-done):

  src/app/validation/names.ts        -> src/shared/validation/names.ts
  src/app/<domain>/validation.ts     -> src/shared/<domain>/validation.ts
  src/app/faq/tags.ts                -> src/shared/faq/tags.ts
  src/app/capture/publisher-*.ts     -> src/shared/asset-publishing/

Adds a `@shared/*` alias for the app side; `convex/` and `workers/` keep
relative imports, which is what they already used for `src/shared` and the
only thing that resolves — the publisher tsconfig has no `paths` block and
there is no convex tsconfig.

The point is what this unlocks. "The server must not import client code"
previously could not be written as a rule, because seven legitimate imports
would have broken it — which is exactly why it existed scoped to one file,
`convex/lib/collaborativeAccess.ts`, where it never fired. With the
exceptions hoisted out, both directions are now general and absolute:
`convex/**` and `workers/**` may not import `src/app/**`. The file-scoped
override keeps only its unrelated trigger-aware-builder rule.

A rule that has to tolerate exceptions cannot be enforced. Move the
exceptions out and then it can — the inverse of widening a guard to fit the
code, which is how the previous boundary ended up with holes.

`publisher-diagnostics.ts` is named in `RENDERER_RUNTIME_CLOSURE_PATHS`, so
the renderer digest changed and `renderer-manifest.generated.ts` is
regenerated here: `code` moved, while `sources`, `toolchain` and `contract`
are byte-identical. That is identity, not staleness — regeneration is driven
by the checked-in `renderer_revisions`, untouched.

Verified: root and publisher typecheck, lint (with probes confirming both new
boundaries reject and that `zod` still passes), format, check:css-orphans,
451 app tests, 181 publisher tests, and publisher:release:verify exiting 0
with the manifest committed.
