+++ .oxlintrc.json
@@ -91,6 +91,28 @@
       }
     },
     {
+      "files": ["convex/**"],
+      "rules": {
+        "no-restricted-imports": [
+          "error",
+          {
+            "patterns": [
+              {
+                "group": [
+                  "**/src/app/**",
+                  "../src/app/**",
+                  "../../src/app/**",
+                  "@app/**",
+                  "@ui/**"
+                ],
+                "message": "The server must not import client code. A contract both sides need — a Zod schema, a tag vocabulary — belongs in `src/shared/`, which both may import."
+              }
+            ]
+          }
+        ]
+      }
+    },
+    {
       "files": ["convex/lib/collaborativeAccess.ts"],
       "rules": {
         "no-restricted-imports": [
@@ -98,13 +120,31 @@
           {
             "patterns": [
               {
-                "group": ["**/src/app/**", "../../src/app/**"],
-                "message": "The server policy module must not import client code (ADR-0001; was #233's allowlisted source-text assertion)."
-              },
-              {
                 "group": ["**/_generated/server"],
                 "importNames": ["mutation", "internalMutation"],
                 "message": "Use the trigger-aware builders from ./functions instead."
+              }
+            ]
+          }
+        ]
+      }
+    },
+    {
+      "files": ["workers/**"],
+      "rules": {
+        "no-restricted-imports": [
+          "error",
+          {
+            "patterns": [
+              {
+                "group": [
+                  "**/src/app/**",
+                  "../src/app/**",
+                  "../../src/app/**",
+                  "@app/**",
+                  "@ui/**"
+                ],
+                "message": "The publisher Worker is a separate deploy artifact and must not import application code. Shared runtime contracts belong in `src/shared/`."
               }
             ]
           }
