+++ src/app/routes/_app/rulesets/$rulesetSlug/faq/create.tsx
@@ -1,4 +1,6 @@
 import { Button, Group, Input, Stack, TextInput, Textarea } from '@mantine/core';
+import type { FaqTag } from '@shared/faq/tags';
+import { FAQ_TAG_LABELS, FAQ_TAG_VALUES } from '@shared/faq/tags';
 import { createFileRoute, Link, useNavigate } from '@tanstack/react-router';
 import { PageLayout } from '@ui/layout/PageLayout';
 import { Surface } from '@ui/surface';
@@ -6,8 +8,6 @@ import { Surface } from '@ui/surface';
 import { useAskFaqQuestion } from '@db/faq';
 import { useCurrentProfile } from '@db/profiles';
 import { loadRulesetBySlug, useRulesetBySlug } from '@db/rulesets';
-import type { FaqTag } from '@app/faq/tags';
-import { FAQ_TAG_LABELS, FAQ_TAG_VALUES } from '@app/faq/tags';
 
 import styles from './create.module.css';
