+++ src/app/rulesets/db.ts
@@ -1,3 +1,4 @@
+import { rulesetInputSchema } from '@shared/rulesets/validation';
 import { useQuery } from 'convex/react';
 import type { FunctionReturnType } from 'convex/server';
 
@@ -5,7 +6,6 @@ import { db } from '@db/core';
 import type { FaqAnswerEntry, FaqItemWithDetails } from '@db/faq';
 import { parseClientBoundary } from '@app/db/core/clientBoundary';
 import { toLiveQueryResult, useLiveMutation } from '@app/db/core/live';
-import { rulesetInputSchema } from '@app/rulesets/validation';
 import { BackgroundClientSchema } from '@game/schema/faction';
 import type { FactionData } from '@game/schema/faction';
