+++ src/app/ui/list/FaqList.tsx
@@ -1,4 +1,6 @@
 import { Anchor, Badge, Group, Stack, Text, Tooltip } from '@mantine/core';
+import { FAQ_TAG_LABELS } from '@shared/faq/tags';
+import type { FaqTag } from '@shared/faq/tags';
 import { Link } from '@tanstack/react-router';
 import { ProfileLink } from '@ui/content/ProfileLink';
 import { SectionedSurface } from '@ui/surface/SectionedSurface';
@@ -7,8 +9,6 @@ import { CircleCheck, CircleDashed } from 'lucide-react';
 import { useMemo } from 'react';
 
 import type { FaqItemWithDetails } from '@db/faq';
-import { FAQ_TAG_LABELS } from '@app/faq/tags';
-import type { FaqTag } from '@app/faq/tags';
 import { formatRelativeDate } from '@app/utils/formatRelativeDate';
 
 import styles from './FaqList.module.css';
