+++ src/shared/asset-publishing/publisher-diagnostics.ts
