+++ src/shared/asset-publishing/publisher-fonts.ts
