+++ src/shared/groups/validation.ts
@@ -1,6 +1,5 @@
+import { alphanumericNameSchema } from '@shared/validation/names';
 import { z } from 'zod';
-
-import { alphanumericNameSchema } from '@app/validation/names';
 
 const groupNameSchema = alphanumericNameSchema('Group name');
