+++ workers/publisher/telemetry.ts
@@ -1,4 +1,4 @@
-import { serializePublisherLogEvent } from '../../src/app/capture/publisher-diagnostics';
+import { serializePublisherLogEvent } from '../../src/shared/asset-publishing/publisher-diagnostics';
 import { rendererManifest } from './renderer-manifest.generated';
 
 export const MAX_TELEMETRY_EVENT_BYTES = 8192;
